rename permissions to management schema
