+++ lib/user-methods.js
@@ -159,7 +159,7 @@ module.exports = {
 
             const realmConstructor = require('./index');
             return new realmConstructor({
-                schema: require('./permissions-schema'),
+                schema: require('./management-schema'),
                 sync: {
                     user: this,
                     url: url.href
